pyproject.toml
[project]
name = "django-hw"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "django>=5.2.9",
    "django-debug-toolbar>=6.1.0",
    "django-environ>=0.12.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djlint>=1.36.4",
    "faker>=39.0.0",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
]
